% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusthistd3.R
\name{clusthistd3}
\alias{clusthistd3}
\title{\code{clusthistd3} This function ...}
\usage{
clusthistd3(
  kk,
  colseq,
  colcol = rep(c(2, 3, 5, 4, 6, 1, 7, 8), 3),
  plot = FALSE
)
}
\description{
\code{clusthistd3} This function ...
}
