% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clusters.R
\name{find_clusters}
\alias{find_clusters}
\title{\code{find_clusters} This function ...}
\usage{
find_clusters(MJS_statistics, distance_threshold, catch_threshold)
}
\description{
\code{find_clusters} This function ...
}
