% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatesamples.R
\name{generatesamples}
\alias{generatesamples}
\title{\code{generatesamples} This function ...}
\usage{
generatesamples(distfunc, densx, trrs, bw = 0, Shift = 0)
}
\description{
\code{generatesamples} This function ...
}
