% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.regionsmm.R
\name{hclust.regionsmm}
\alias{hclust.regionsmm}
\title{\code{hclust.regionsmm} This function ...}
\usage{
hclust.regionsmm(distribmatin, rr, adj = FALSE, adjmat = NULL, KL = "KL")
}
\description{
\code{hclust.regionsmm} This function ...
}
