% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterodists.R
\name{heterodists}
\alias{heterodists}
\title{\code{heterodists} This function ...}
\usage{
heterodists(
  merges,
  distseq,
  densx,
  densmaty,
  rrs,
  KL = "KL",
  BB = c(100, 1000),
  bounds = c(0.005, 0.1)
)
}
\description{
\code{heterodists} This function ...
}
