% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatesample.R
\name{generatesample}
\alias{generatesample}
\title{\code{generatesample} This function ...}
\usage{
generatesample(distfunc, densx, trrs, nsize = 50, bw, Shift = 0)
}
\description{
\code{generatesample} This function ...
}
