% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterodist.R
\name{heterodist}
\alias{heterodist}
\title{\code{heterodist} This function ...}
\usage{
heterodist(
  merges,
  distseq,
  densx,
  densmaty,
  rrs,
  KL = "KL",
  doko = c(1, 1),
  BB = 5,
  bw = 0.1
)
}
\description{
\code{heterodist} This function ...
}
